Seprate IsServer and IsClient into modules
This will help us write tests for each without making the module too big with similar function names in the same module
